Created Customer Controller, edited delete from CustomerService to return boolean

+++ backend-api/src/main/java/com/CSC340/MinervasList/service/CustomerService.java
@@ -38,7 +38,12 @@ public class CustomerService {
         return customerRepository.save(existingCustomer);
     }
 
-    public void deleteCustomer(Long id) {
-        customerRepository.deleteById(id);
+   public boolean deleteCustomer(Long id) {
+    if (!customerRepository.existsById(id)) {
+        return false;
     }
+
+    customerRepository.deleteById(id);
+    return true;
+}
 }
